+++ inst/examples/germany_missing.R
@@ -4,14 +4,14 @@ library(data.table)
 library(purrr)
 library(ggplot2)
 
-# Use 2 cores
-options(mc.cores = 2)
+# Use 4 cores
+options(mc.cores = 4)
 
 # Load and filter germany hospitalisations
 nat_germany_hosp <- germany_covid19_hosp[location == "DE"][age_group %in% "00+"]
 nat_germany_hosp <- enw_filter_report_dates(
   nat_germany_hosp,
-  latest_date = "2021-10-01"
+  latest_date = "2021-08-01"
 )
 
 # Make sure observations are complete
@@ -21,46 +21,16 @@ nat_germany_hosp <- enw_complete_dates(
   missing_reference = FALSE
 )
 
-# Prototypes for simulating missing data - likely to be implemented in 0.2.0
-enw_incidence_to_cumulative <- function(obs, by = c()) {
-  obs <- data.table::as.data.table(obs)
-  obs <- check_dates(obs)
+# Set proportion missing at 35%
+prop_miss <- 0.35
 
-  obs <- obs[!is.na(reference_date)]
-  obs[order(reference_date, report_date)]
-
-  obs[, confirm := cumsum(new_confirm), by = c(by, "reference_date")]
-  return(obs[])
-}
-
-enw_simulate_missing_reference <- function(obs, proportion = 0.2, by = c()) {
-  obs <- data.table::as.data.table(obs)
-  obs <- check_dates(obs)
-  obs <- enw_assign_group(obs, by = by)
-  by_with_group_id <- c(".group", by)
-  obs <- enw_new_reports(obs)
-
-  obs[, missing := purrr::map2_dbl(
-    new_confirm, proportion, ~ rbinom(1, .x, .y)
-  )]
-  obs[, new_confirm := new_confirm - missing]
-
-  complete_ref <- enw_incidence_to_cumulative(obs, by = by)
-  complete_ref[, c("new_confirm", ".group", "delay", "missing") := NULL]
-
-  missing_ref <- obs[, .(confirm = sum(missing)),
-    by = c(by, "report_date")
-  ]
-  missing_ref[, reference_date := as.IDate(NA)]
-
-  obs <- rbind(complete_ref, missing_ref, use.names = TRUE)
-  obs[order(reference_date, report_date)]
-  return(obs[])
-}
-
+<<<<<<< HEAD
 # Simulated proportion missing
 prop_miss <- 0.6
 
+=======
+# Simulate using this function
+>>>>>>> feature-missing-reference-function
 nat_germany_hosp <- enw_simulate_missing_reference(
   nat_germany_hosp,
   proportion = prop_miss, by = c("location", "age_group")
@@ -73,32 +43,37 @@ retro_nat_germany <- enw_filter_report_dates(
 )
 retro_nat_germany <- enw_filter_reference_dates(
   retro_nat_germany,
-  include_days = 40
+  include_days = 60
 )
 
 # Get latest observations for the same time period
 latest_obs <- enw_latest_data(nat_germany_hosp)
 latest_obs <- enw_filter_reference_dates(
   latest_obs,
-  remove_days = 40, include_days = 20
+  remove_days = 40, include_days = 60
 )
 
 # Preprocess observations (note this maximum delay is likely too short)
 pobs <- enw_preprocess_data(retro_nat_germany, max_delay = 20)
 
-# Compile nowcasting model using multi-threading
-model <- enw_model(threads = TRUE)
+# Compile nowcasting model without multi-threading as only using a single
+# group and the missing reference only supports multi-threading across groups
+model <- enw_model(threads = FALSE)
 
 # Fit the nowcast model with support for observations with missing reference
 # dates and produce a nowcast
 # Note that we have reduced samples for this example to reduce runtimes
 nowcast <- epinowcast(pobs,
+<<<<<<< HEAD
   missing = enw_missing(~ 1 + (1 | day), data = pobs),
+=======
+  missing = enw_missing(~ (1 | week), data = pobs),
+>>>>>>> feature-missing-reference-function
   report = enw_report(~ (1 | day_of_week), data = pobs),
   fit = enw_fit_opts(
     save_warmup = FALSE, pp = TRUE,
-    chains = 2, threads_per_chain = 2,
-    iter_warmup = 500, iter_sampling = 500
+    chains = 4, iter_warmup = 500, iter_sampling = 500,
+    likelihood_aggregation = "groups", adapt_delta = 0.9
   ),
   obs = enw_obs(family = "poisson", data = pobs),
   model = model
@@ -118,7 +93,7 @@ miss_prop <- cbind(
 ggplot(miss_prop) +
   aes(x = reference_date) +
   geom_line(aes(y = median), size = 1, alpha = 0.6) +
-  geom_line(aes(y = mean), linetype = 2) +
+  geom_hline(yintercept = prop_miss, linetype = 2) +
   geom_ribbon(aes(ymin = q5, ymax = q95), alpha = 0.2, size = 0.2) +
   geom_ribbon(aes(ymin = q20, ymax = q80, col = NULL), alpha = 0.2) +
   geom_hline(yintercept = prop_miss, linetype = 2) +