+++ tests/testthat/test-enw_complete_dates.R
@@ -15,14 +15,18 @@ test_that("enw_complete_dates works as expected with well behaved data", {
     ),
     confirm = c(rep(1, 3), rep(0, 6))
   )
+  data.table::setkeyv(exp_obs, c("reference_date", "report_date"))
   expect_equal(enw_complete_dates(obs), exp_obs)
   expect_equal(
     enw_complete_dates(obs, missing_reference = FALSE),
     exp_obs[!is.na(reference_date)]
   )
   obs$location <- "DE"
+  data.table::setkeyv(
+    exp_obs[, location := "DE"], c("location", "reference_date", "report_date")
+  )
+  data.table::setcolorder(exp_obs, "location")
   expect_equal(
-    enw_complete_dates(obs, by = "location"),
-    data.table::setcolorder(exp_obs[, location := "DE"], "location")
+    enw_complete_dates(obs, by = "location"), exp_obs
   )
 })